refactor(uv): extract lock index credential matcher
