+++ uv/lib/dependabot/uv/file_updater/lock_file_updater.rb
@@ -3,7 +3,6 @@
 
 require "toml-rb"
 require "open3"
-require "uri"
 require "dependabot/dependency"
 require "dependabot/shared_helpers"
 require "dependabot/uv/language_version_manager"
@@ -25,6 +24,7 @@ module Dependabot
         require_relative "pyproject_preparer"
         require_relative "version_config_parser"
         require_relative "lock_file_error_handler"
+        require_relative "lock_index_credential_matcher"
 
         REQUIRED_FILES = %w(pyproject.toml uv.lock).freeze # At least one of these files should be present
 
@@ -364,9 +364,10 @@ module Dependabot
 
           options = T.let([], T::Array[String])
           used_credential_urls = T.let([], T::Array[String])
+          credential_matcher = LockIndexCredentialMatcher.new(credentials: filtered_credentials)
 
           uv_lock_registry_urls.each do |registry_url|
-            credential = best_credential_for_registry_url(filtered_credentials, registry_url)
+            credential = credential_matcher.best_credential_for_registry_url(registry_url)
             next unless credential
 
             used_credential_urls << credential["index-url"].to_s
@@ -397,56 +398,6 @@ module Dependabot
         def authed_registry_url(credential, registry_url)
           lock_credential = Dependabot::Credential.new(credential.to_h.merge("index-url" => registry_url))
           AuthedUrlBuilder.authed_url(credential: lock_credential)
-        end
-
-        sig do
-          params(credentials: T::Array[Dependabot::Credential], registry_url: String)
-            .returns(T.nilable(Dependabot::Credential))
-        end
-        def best_credential_for_registry_url(credentials, registry_url)
-          credential_scores = credentials.map do |credential|
-            [credential, credential_match_score(credential["index-url"].to_s, registry_url)]
-          end
-          best_match = credential_scores.max_by { |_, score| score }
-
-          return nil unless best_match
-          return nil if best_match[1].negative?
-
-          best_match[0]
-        end
-
-        sig { params(credential_url: String, registry_url: String).returns(Integer) }
-        def credential_match_score(credential_url, registry_url)
-          normalized_credential_url = normalize_index_url(credential_url)
-          normalized_registry_url = normalize_index_url(registry_url)
-
-          return 100_000 if normalized_credential_url == normalized_registry_url
-
-          credential_uri = URI.parse(normalized_credential_url)
-          registry_uri = URI.parse(normalized_registry_url)
-
-          return -1 unless credential_uri.scheme == registry_uri.scheme
-          return -1 unless credential_uri.host == registry_uri.host
-          return -1 unless credential_uri.port == registry_uri.port
-
-          credential_path = normalized_uri_path(credential_uri)
-          registry_path = normalized_uri_path(registry_uri)
-
-          return 1 if credential_path == "/"
-
-          if registry_path.start_with?(credential_path.chomp("/") + "/")
-            credential_path.length
-          else
-            -1
-          end
-        rescue URI::InvalidURIError
-          -1
-        end
-
-        sig { params(uri: URI::Generic).returns(String) }
-        def normalized_uri_path(uri)
-          path = uri.path.to_s
-          path.empty? ? "/" : path
         end
 
         sig { returns(T::Array[String]) }
